# docs/openapi-schema.json
openapi: 3.0.3
info:
  title: VORA Platform API
  version: 1.0.0
  description: "\n# VORA — Plateforme Taxi Jaune (Cameroun)\n\nAPI REST unifiée pour
    l'écosystème **VORA** : application passager, cockpit chauffeur (taxi jaune),\ngéolocalisation
    PostGIS, optimisation de tournées, paiements Mobile Money simulés et administration.\n\n**Production
    :** `https://vora-ujbv.onrender.com/api/v1`  \n**Documentation interactive :**
    `/api/v1/docs/`  \n**Schéma OpenAPI :** `/api/v1/schema/`\n\n---\n\n## Architecture
    système\n\n| Couche | Rôle |\n|--------|------|\n| **Frontend** (mobile / web)
    | UI passager & chauffeur — consomme cette API |\n| **Django (ce service)** |
    Auth JWT, courses, paiements simulés, zones/carrefours, trajectoires, optimisation
    PostGIS |\n| **PostgreSQL + PostGIS** | Données géo (zones, carrefours, corridors,
    positions chauffeurs) |\n| **Google Maps API** | Géocodage, itinéraires, congestion
    |\n| **Spring Boot** *(optionnel)* | Réservations — appelle `verify-destination`
    et `optimize/turn` |\n\n---\n\n## Rôles utilisateur\n\n| Rôle | `role` JWT | Espace
    frontend | Endpoints principaux |\n|------|------------|-----------------|----------------------|\n|
    **Passager** | `passenger` | `pickup` — commander une course | `/rides/*`, `/payments/*`,
    `/safety/sos` |\n| **Chauffeur** | `driver` | `driver-cockpit` — habitacle & revenus
    | `/driver/*`, `/drivers/online` (lecture) |\n| **Admin** | `admin` *(futur)*
...
paths:
...
  /api/v1/carrefours:
    get:
      operationId: carrefours_list
      parameters:
      - in: query
        name: city
        schema:
          type: string
      - in: query
        name: zone_id
        schema:
          type: integer
      tags:
      - Carrefours
      security:
      - BearerAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Carrefour'
          description: ''
    post:
      operationId: carrefours_create
      tags:
      - Carrefours
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarrefourCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CarrefourCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CarrefourCreateRequest'
        required: true
      security:
      - BearerAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Carrefour'
          description: ''
...
components:
  schemas:
...
    Carrefour:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        zone_id:
          type: integer
          nullable: true
          readOnly: true
        name:
          type: string
          maxLength: 150
        city:
          type: string
          maxLength: 50
        location:
          type: string
          readOnly: true
        is_pickup_point:
          type: boolean
        is_major_intersection:
          type: boolean
        typical_waiting_passengers:
          type: integer
          maximum: 32767
          minimum: 0
        standard_fare_to_centre_fcfa:
          type: integer
          maximum: 2147483647
          minimum: 0
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - location
      - name
      - updated_at
      - zone_id
    CarrefourCreateRequest:
      type: object
      properties:
        zone_id:
          type: integer
          nullable: true
        name:
          type: string
          minLength: 1
          maxLength: 150
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        is_pickup_point:
          type: boolean
          default: true
      required:
      - lat
      - lng
      - name
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT obtenu via `POST /auth/signin` ou `POST /auth/otp/verify`.
        Format header : `Authorization: Bearer <access_token>` — validité 24 h.'
servers:
...